Add validation of --options passed to commands
Validate arguments passed to command line. The execution stops when the argument is not part of known options or if it requires value, but the user had not passed one.
Add tests for empty string value of known option and value with spaces only. This way we verify that even if the passed value to known option is valid string, but its value after trimming is empty, we will not execute the command.
Use osenv for defaultProfileDir as it is platform independent. Add fail message when directory passed to init command does not exist. Add validateYargsArguments tests.
Use optional parameter of validateYargsArguments - isInTestMode, to trigger validation for passed options. Remove process.argv[1] changes from the tests, as they could cause unexpected behavior.

http://teampulse.telerik.com/view#item/262916
http://teampulse.telerik.com/view#item/250186