Improve remote message
Add more information for the users how to use remote command. The information is printed when the command is executed. This will make it easier for them to understand where to use the port number, which is passed as parameter to this command.
Add DomainNameSystem class to be used to reverse ip address to DNS. Add ip module and use it to show the exact ip address that the user should use.

http://teampulse.telerik.com/view#item/274327 (+1 squashed commits)