Fix android emulator when more than one device is used
Remove passing of options.avd to androidEmulatorService as this is a common option and it can be used there directly. AndroidEmulatorService is also part of the lib, so there's no need to pass as argument something that it can easily accessed.
Add help for retryCount option which can be used in emulate android command.

http://teampulse.telerik.com/view#item/277469