Release console when ios-sim cannot start
Update common lib where the following changes are applied:
Add call for ios-sim --version in order to determine if Xcode and ios-sim versions are compatible. In case they are not, the command will return error - catch it and break the whole execution. Print the error message from ios-sim --version, but remove its stack first as any throw in child_process.stderr.on(...) prints its callstack automatically. So new option is added to errors - hideCallStack. In case it is true, the stack option is set to null.