Update fibers dependencies
Update fibers and mocha-fibers dependencies in package.json to use release branch instead of master. This way we can make updates to master branches during our development. Currently release branches of fibers and mocha-fibers are exact copies of their master branches, so this change should not make any change to the functionality.