Replace retryCount with timeout
Move timeout option to common options. Update common lib where the following change is applied:
When starting Android emulator we were using --retryCount option, which is difficult to be explained. It it replaced with --timeout option, which is the allowed time to start the emulator device(wait to boot). The default value is 120 seconds. retryCount option is removed as it is not used anymore.
Update help and error messages.

http://teampulse.telerik.com/view#item/278845