Appstore commands prompt for AppleID
If you do not specify AppleID, appstore commands prompt you to provide it instead of failing with error. Add AppstoreApplicationCommandBase and change current appstore commands to extend it. Add getAppleId method in order to get the id. Change in help.txt to show that AppleID is no longer mandatory parameter.

http://teampulse.telerik.com/view#item/274748