Fix --save-to option
We are validating all -- options. Options which have dash ("-") are added by yargs in two ways, in this case --save-to is passed to our code as save-to and saveTo. As we have not added saveTo to known options, trying to use --save-to fails. Added saveTo option to knownOpts. Added test that verifies all options, which have "-" are added correctly to options (in both ways).