Fix StringCommandParameter
When injector tries to resolve StringCommandParamater, it raises error as it's unable to use the current constructor. Add StringParameterBuilder class that is used to generate string parameters. Add createMandatoryParameter method to it, so it will generate mandatory parameter with specified error message. Add IStringParameterBuilder interface.
The idea is to use stringParameter from injector resolver when the parameter will not be mandatory and to use the builder when a more complicated logic is needed.

Fixes http://teampulse.telerik.com/view#item/279466