Remove # from listing
Remove # sign when listing available certificates, provisions, cloud projects and devices (appbuilder certificate, appbuilder provision, appbuilder cloud and appbuilder device commands).

Currently some of our commands print lists and place numbers at the beginning of each row, for example:
#1: "Project 1"
#2: "Project 2"

This leads to confusion when passing arguments to commands like --certificate, where it is expected to pass the number (1), but some customers pass #1. In shell scripts # is sign for comment, so everything after it is ignored and not passed as command parameter, which leads to incorrect behavior. So remove the # in order to make it easier for the customers to understand that we require the number (or the name).