Fix plugins commands
Change platforms property of IPlugin from string[] to Server.DevicePlatform[] type. Add try-catch when fetching data for specific plugin - this way if we fail while fetching data for one of the plugins, information about rest of them will be shown to user. Add test to verify plugins data will be shown correctly when we are unable to get data for one of the plugins.

Change url to MarketPlace to use https instead of http.

Fixes http://teampulse.telerik.com/view#item/280934
http://teampulse.telerik.com/view#item/280394