Appstore commands check if you are logged in
When user is not logged in our platform Appstore commands prompt him for Apple ID and password and after that show message "Not logged in". This is confusing as it seems apple credentials are incorrect. Add login page prompt before checking the id and password.

Fixes http://teampulse.telerik.com/view#item/280852