Windows Phone SDK updates
Add check for current Windows Phone SDK in .abproject (WPSdk) and when migrating it (using prop set) check if the new value is supported in the current Cordova version. As WP 8.1 SDK is supported in Cordova 3.7.0 and above, if the project is trying to use WP 8.1 SDK and its version is 3.5.0 or below, a prompt message suggests to use newer Cordova version. In case user is migrating his Framework version to a lower than 3.7.0 we check WPSdk value and if it is 8.1, we prompt the user to use 8.0.
Update common lib where new method has been added to Errors: failWithoutHelp.

Fix unit tests by adding WPSdk to .abproject in test/resources Cordova file.

http://teampulse.telerik.com/view#item/284537