Add check if 7za can be executed on Linux
If user is on x64 Linux and 32 bit libraries, describe in our documentation, are not installed, user will not be able to use 7zip. This means - unable to create projects, unable to build etc. So on post-install we check if we are able to execute the 7za file and if ENOENT is return we show the user where he can find system requirements.

Fix host-info paths in some files and remove host-info reference from some other files.

Remove duplicated methods from host-info files (they exist on two places - in common lib and here).

http://teampulse.telerik.com/view#item/285027