Add --download option to appmanager upload
If you are using Development provision and you execute $appbuilder appmanager upload ios, the .ipa is downloaded. But if your provision is any other type or you upload for any other platform, the project is not downloaded. Change build code to download project with development certificate only if it is not build for appmanager. Use --download option with appmanger upload command in order to download the file.

http://teampulse.telerik.com/view#item/285204