Fix emulate of NativeScript projects
Fix getProjectTargets of nativescript-project to be "android" and "ios" (remove code which was checking for bootstrap.android.js and bootstrap.ios.js). Fix emulate command to use appIdentifier directly (without using AppIdentifier class and passing incorrect options.companion option to it). Remove projectTargets property of project, as there is getProjectTargets method, which is used almost everywhere when we need project targets (fixed the only place where the projectTargets property was used).  Add emulate option to project capabilities in order to forbid emulate of mobile-website projects. EmulatorSettingsService checks for emulate capability and fails in case the project cannot be emulated.

NOTE: This PR allows to deploy nativescript projects to native emulator. For Android, the application is installed, but it is not started. It has to be started manually. Graphite is doing the same, I'll research the issue and if I find a fix, I'll include it in another PR.

Fixes http://teampulse.telerik.com/view#item/285409