Update fibers and ios-sim usage
Update package.json to use fibers version 1.0.5.1 (it is the same as the currently used from branch vladimirov/fibers_1_0_5_fixes.
Use ios-sim-portable version 1.0.2 - it is the same as the master branch.
Use correct syntax to check if version is below 0.11.0 (if 0.11.00 is used, warning is printed even if your version is above 0.10.35 and below 0.11).

IMPORTANT: This pull request can be merged ONLY AFTER we publish version 1.0.2 of ios-sim-portable.