Forbid ports below 1024
With latest XCode, when you try starting simulator with sudo, the application is not started. Currently we are using sudo when we want user wants to use port below 1024. Even if he uses $ sudo appbuilder remote 444, the application will not start because of the bug in XCode. So using ports below 1024 is useless. That's why we forbid using them.
Fixes http://teampulse.telerik.com/view#item/285912