Remove duplicate -- options
Yargs adds -- options to yargs.argv in two ways (--profile-dir is added as profile-dir and profileDir). In order to workaround this behavior, we had to add  all dashed options in both ways to our known options. Fix our validation logic, so we'll have to add only profile-dir.
Remove check for client name when validating -- options. This way options will be validated no matter if CLIENT_NAME( tns) or CLIENT_NAME_ALIAS (nativescript) is used on command line.

Remove options.ts tests as it is not needed anymore. Add tests in common-helpers.