Fix marked-terminal parsing and tables in console
Remove the hack when parsing markdown text that has to be show on the console. The problem was with code in table, but it had been fixed in version 1.3.0 of marked-terminal.
Use telerik fork of cli-table in order to show tables correctly.

Part of http://teampulse.telerik.com/view#item/259417