Fix installation and uninstall of CLI
Fix installation of appbuilder-cli by removing ProjectTemplates, ItemTemplates, etc. directories from .npmignore file. This way our resources are uploaded to npm and we are working fine.
Fix preuninstall script which was failing on several places (missing lodash as it was not required, unable to resolve injector, etc.) In order to fix the issues, new command  "dev-preuninstall" is introduced. This way injector can be resolved.