AppManager LiveSync
Allow publishing of patches in AppManager. Add new command: `appmanager livesync`. It can be used without parameters in which case the user must confirm that the patch should be published to all platforms. The command accepts multiple arguments, which are the platforms. It can be used with one or with many platforms:
* `$ appmanager livesync android`
* `$ appmanager livesync android ios`
etc.

In case LivePatch plugin is not enabled for release configuration, we notify the user the we are enabling it.
Async plugins service gathering of plugin information.
http://teampulse.telerik.com/view#/12319