Fix plugin add/remove
When plugin is enabled in only one configuration (for example in release only), trying to remove it with `$ appbuilder plugin remove <pluginid>` is telling you that the plugin is not enabled for this project.
When plugin is enabled in only one configuration (for example in release only) trying to add it with `$ appbuilder plugin add <pluginid>` is adding it to both configurations, so in release config the plugin is added twice.
When plugin is enabled in only one configuration (for example in release only) trying to add it with `$ appbuilder plugin add <pluginid> --debug --release` is adding it to both configurations, so in release config the plugin is added twice.

Remove `getInstalledPluginsEnabledAtLeastInOneConfiguration` and use its logic as default for getInstalledPlugins. This way we always get full list of plugins enabled in all configurations. In case you specify configuration (`--release` for example) `getInstalledPlugins` will give you only the plugins enabled in this configuration.

Fixes http://teampulse.telerik.com/view#item/290735