Fix marketplace plugins commands
Make sure plugin with different versions is not enabled in different configurations (ask the user to select which version to use in this case - list all available versions).
Ask the user for confirmation if specified version should be enabled in one config only and removed from the other.
Add unit tests