Prevent doubling of typescript definitions
CLI has its own typescript definition files, which are used when we build typescript projects. Make sure to exclude them in case the project directory already has files with such name. This is the scenario with NativeScript TypeScript projects, which have android17.d.ts files in the templates and we are adding them again on build.