Improve update-kendoui command
Improve update-kendoui command to accept several -- options:
* `--verified` - lists verified plugins
* `--core` - lists core plugins
* `--professional` - lists professional plugins

They can be combined, but core and professional cannot be used simultaneously. For example `$ appbuilder update-kendoui --core --verified` will list only verified core plugins.

Add `--latest` option. If it is specified, the user will not be prompted to select version, but the latest available version will be installed. This option can be combined with the other options. for example:
* `$ appbuilder update-kendoui --verified --latest` will install latest available verified version
* `$ appbuilder update-kendoui --verified --core --latest` will install latest available verified core version

http://teampulse.telerik.com/view#item/288592
http://teampulse.telerik.com/view#item/284131