Fix update-kendoui --professional --latest
When user cannot use professional versions and tries to execute `appbuilder update-kendoui --professional --latest` leads to `Cannot read property 'DownloadUrl' of undefined`
The problem is that packages collection that we are using becomes empty and we are trying to get the first item of it (undefined), so it does not have DownloadUrl property.
Prevent this case by checking the length of packages collection and show correct error message in case --professional is specified.

Fixes http://teampulse.telerik.com/view#item/291742