Fix upload of nativescript projects to appmanager
Fix upload of nativescript projects to appmanager. The issue is raised from resolving `pluginsService` in the constructor of AppmanagerService. PluginsService constructor is failing when the project is Cordova.
Use injector to resolve pluginsService only where it is required.
Fix error when plugin is enabled manually in configuration, but we are unable to find it in marketplace (for example manually add com.telerik.LivePatch in release config and try $ appbuilder appmanager livesync android with user who does not have access to this feature. You'll receive error **Cannot read property data of undefined**

Fixes http://teampulse.telerik.com/view#item/291755