When the project is migrated we have to update tns_modules directory and .abproject with the new version of the project. Add new NativeScriptMigrationService and use it with dynamic call to upgrade the project. In case an error occurs during migration, rollback the migration. In case current project version is not supported, we have to break the migration. Oboslete versions can still be build, but we cannot migrate them. Download NativeScript migration data during prepublish. Make sure to save supportedVersion in format - DisplayName(string) and Version(string). You can change the version by using: `$ appbuilder prop set FrameworkVersion 1.0.1` (or 0.10.0) or `$ appbuilder mobileFramework set 1.0.1` You can list valid versions by using `$ appbuilder prop print FramworkVersion --validValue` or `$ appbuilder mobileFramework`
Add two new interfaces, one for NativeScriptMigration data and one describing Framework versions (with version and display name). They second one is used to unify the way we get supported framework versions in Cordova and NativeScript projects. Until now, we were calling the server to get the Cordova supported framework versions. In fact we can have this information stored in cordova-migration-data.json. Store it there and when we want to check the data, read the file. The same behavior is used for NativeScript.

Implements http://teampulse.telerik.com/view#item/292172