Add latest or default plugin versions
Add support for installing the latest plugin version. The command is: `$ appbuilder plugin add toast --latest`. This will install the latest available version of Toast plugin.
Add support for installing the default plugin version. The command is: `$ appbuilder plugin add toast --default`. This will install the default version of Toast plugin.
As on the server MarketplacePluginVersionsData has three more properties, which are not generated in our server-api, implement IMarketplacePluginVersionsData interface, that extends this class and add the properties there.

http://teampulse.telerik.com/view#item/292306