Set Description when creating new project
When a new Cordova project is created, we have to set its Description to be equal to projectName. This is due to this change: https://github.com/Icenium/Ice/pull/3729
NOTE: The change on the server reflects only Cordova projects, but it will be changed to update NativeScript project's descriptions as well.
This will fix our unit tests.