Pass CordovaPluginVariables on the command line
When you want to automate plugin add, there's no way to automate the process of adding CordovaPluginVariables and prompter is shown for each of them. In order to simplify this, add `var` dashed option, which should work in the following way:
* `$ appbuilder plugin add dropbox --var.APP_KEY 1 --var.APP_SECRET 2` - sets the plugin variables APP_KEY to 1 and APP_SECRET to 2 in all configurations
* `$ appbuilder plugin add dropbox --var.debug.APP_KEY 1 --var.debug.APP_SECRET 2 --var.release.APP_KEY 3 --var.release.APP_SECRET 4` - in .debug.abproject sets APP_KEY to 1 and APP_SECRET to 2 and in .release.abproject sets APP_KEY to 3 and APP_SECRET to 4
* `$ appbuilder plugin add dropbox --var.APP_KEY 1 --var.APP_SECRET 2  --var.debug.APP_KEY 3` - in .debug.abproject sets APP_KEY to 3 and APP_SECRET to 2 and in .release.abproject sets APP_KEY to 1 and APP_SECRET to 4 - when plugin variable is passed both with config and without config, the value for config is used for the respective configuration, the other value is used for all other values.
* `$ appbuilder plugin add dropbox --var.APP_KEY 1` - sets APP_KEY to 1 in all configurations, and prompts user for APP_SECRET for all configurations
* `$ appbuilder plugin add dropbox --var.APP_KEY 1 --var.debug.APP_SECRET 2` - sets APP_KEY to 1 in all configurations, APP_SECRET to 2 in debug config and prompts user for APP_SECRET value for release configuration

Also keys passed to --var will work no matter of the casing, for example --var.DeBuG.App_KeY is treated the same way as --var.debug.APP_KEY and will be used to set APP_KEY in debug configuration to passed value.

Added unit tests for the described scenarios.

http://teampulse.telerik.com/view#item/292387