Fail in prompter when console is not interactive
There's no point in using the prompter when the console is not interactive.
CLI should still be usable even if the console is not interactive so we should provide a way for every command to bypass using the prompter one way or another.
Part of [#297090](http://teampulse.telerik.com/view#item/297090)'s implementation