Fix error message of build commands
When build command is executed outside of project dir, its execute logic is triggered and the hook that's attached to it is throwing an error. This error is caught by our uncaughtException listener, which prints message with call stack and exits the process. We have catch for this code, but it looks like the problem is with the two fibers working in parallel.
Also update common  lib, where the following change is applied:

In case command does not have default command and arguments are not passed, we are trying to execute default command, which is not expected. Make sure not to return true in this method, when there's no default command.

Fixes: http://teampulse.telerik.com/view#item/297504