Allow export of apps and solutions using new server API
Due to changes in Server API, we have to allow exporting of apps and solutions.
In order to get projects from apps, we should call TAP endpoint.
For not migrated projects we should call the old server endpoint.
The export procedure requires calling same endpoint for migrated and not migrated projects.
When exporting old projects (from solutions), we should pass the solution name as appId and send the magic Header in the request.
When exporting new projects (from apps), we should pass real appId and DO NOT send the magic Header.
In order to achieve this, introduce new class - ServiceProxyBase and use it for TAP requests.
Introduce AppBuilderServiceProxy class that extends ServiceProxyBase and register it as "serviceProxy", so we'll keep current behavior for server calls.
Add new method in AppBuilderServiceProxy that allows sending server requests with passing "X-Icenium-SolutionSpace" header.
Add "NOT MIGRATED" string to solutions when they are listed.