Start android applications after deploy
Fixes http://teampulse.telerik.com/view#item/306407

Update common lib, where the following change is applied:
Add canStartApplication to applicationManager

We have different cases when we can start the application, based on current platform and device platform.
Add `canStartApplication` in applicationManager's which evaluates all circumstances and can be used to check if we are able to start the application on device.