Print information about plugin variables
Print information about nativescript plugin variables.
Print the values when plugin is already added (`appbuilder plugin` command and when lisiting available plugins: `appbuilder plugin --available`).
Print plugins variables when listing available plugins (valid for Cordova and {N} Plugins) - even when plugin is not added, we know which are the required varialbes, show them to the user.