Fix plugins list for obosleted plugins
When we have obsoleted plugin with the same identifier as its marketplace replacement, current code filters it.
For example in case you add `com.telerik.feedback` and execute `appbuilder plugin`, it will not be listed.
Filter only items where the replacement id is not the same as integrated plugin's id.

Set version to 3.1.0