Fix overwriting DisplayName when iOSDisplayName is present and empty
If `iOSDisplayName` is present but an empty string, then our current logic sets the `DisplayName` to `iOSDisplayName` and right after that decides that no `DisplayName` is present and uses a default one.
Fix the issue, by checking if a `DisplayName` is already set and then fork the logic between `iOSDisplayName` and the default one.