Fix plugin configure command
When the .debug.abproject and the .release.abproject files are deleted and then plugin is installed without specified configuration and the plugin variables are set in the .abproject file. When configuring the plugin variables for specific configuration the original plugin variables must be cloned because the changes must be saved in the configuration specific file only.
When getting the plugin vatiable from var option the configuration needs to be checked because if there is no specified configuration in the command it will be undefined.