Integrate Ionic CLI projects with AppBuilder CLI
The projects created with Ionic CLI have Ionic specific srtucture and some changes should be done to transform the project to AppBuilder hybrid project.
Rerouting index.html file should be created to pointto the Ionic index.html file in the www folder.
The resources should be copied from the resource folder to App_Resouces folder. Android resources should be renamed and put to the appropriate folders. Windows Phone 8 splashscreen should be JPG not PNG.
The information in the config.xml of the Ionic project should be split and moved to the appropriate AppBuilder folders. Also the sources of the resources should be changed to be relative to the newly added platform specifix config.xml.
All AppBuilder default plugins should be removed from the plugins folder to reduce the size of the project.
All unnecessary files and folders from the Ionic project should be removed.