Search http://registry.npmjs.org for plugins
When existing plugin name is passed to the fetch command it takes very long time to fetch the plugin. Slow fetch is expected only when keywords are passed to the fetch command. We can send request to http://registry.npmjs.org/[plugin-name] and see if there is registered plugin before using the slow npm find command.