Download build results from s3
If we want to download the build results directly from s3 we need to send one additional property in the request - AcceptResults, with value - Url;LocalPath. In the response the link to the result will be in the FullPath property and the Format will be Url. In case we can't download the build results from s3 (wp8) we need to use the LocalPath method.