Respect configurations on appmanager/appstore upload
Expose a way to default to a desired configuration (if present among the user configurations) throught the `$project` dependency.
Set project configuration as:
* What the user has defined as `--config`
* `release` (if present among the user's configurations)
* The lexicographically first configuration the user has
* `Debug`

In that order.

Mind that this is a project configuration as opposed to build configuration and even in the unlikely case that we end up building in `Debug`, the build configuration will always be `release` thus preventing any unwanted behavior.