Install all npm dependencies in NS TS projects
When we create NS TS project we need to install all npm dependencies because if we don't have the tns-core-modules installed we will not add theirs .d.ts in .abreferences file untill the project is built.
Also we need to install the dependencies on every project related command because we must be sure that we are working with correct content of node_modules.