Validate hook arguments before execute
Since we can have hooks with arguments which are invalid for the AppBuilder CLI (for example the before-livesync hook of nativescript-angular and watch hook of nativescript-dev-typescript) we need to check if the arguments of the hook are valid before we execute it. We can do the check only for hooks which can be executed in our process.