Fix adding of NativeScript plugins
Fix code that verifies if NativeScript plugin can be used in the project. Each nativescript plugin may describe the minimum required version of the framework. Our check is not correct and it does not allow adding of nativescript-plugin-livepatch plugin (it's required versions are 1.x.x).
Fix the check - now if all versions described in plugin's package.json are greater than current framework version, the plugin will not be added. In case at least one of the version described in plugin's package.json is compatible with our FrameworkVersion in .abproject, the plugin will be added.